src/autism_project/requirements.txt
# Python dependencies for Auricare Chatbot
python-dotenv==1.0.0
pandas==2.2.3
requests==2.31.0
pypdf==3.17.4
groq==0.4.1

# LangChain ecosystem (compatible set)
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.52
langsmith==0.1.17
langchain-groq==0.1.3
langgraph==0.0.20

# Vector search
faiss-cpu==1.12.0
sentence-transformers==2.2.2

# Frontend / API
gradio==4.44.0
flask==3.0.0
flask-cors==4.0.0
